@@ -4,7 +4,8 @@
   "description": "Mutualized toolchain & foundation for react application",
   "main": "./src/index.js",
   "scripts": {
-    "test": "./node_modules/eslint/bin/eslint.js ./src"
+    "test": "npm run lint",
+    "lint": "./node_modules/.bin/eslint ./src"
   },
   "keywords": [
     "toolchain",
@@ -14,70 +15,40 @@
   "author": "Johan Girod <johan@evaneos.com>",
   "license": "UNLICENSED",
   "devDependencies": {
-    "babel-core": "^6.4.0",
-    "babel-loader": "^6.2.1",
+    "babel-core": "^6.4.5",
+    "babel-eslint": "^5.0.0",
+    "babel-loader": "^6.2.2",
     "babel-plugin-react-require": "^2.1.0",
-    "babel-plugin-react-transform": "^2.0.0",
-    "babel-plugin-transform-class-properties": "^6.4.0",
-    "babel-plugin-transform-object-rest-spread": "^6.3.13",
-    "babel-polyfill": "^6.3.14",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-preset-es2015-node": "^5.0.2",
+    "babel-plugin-typecheck": "^3.6.1",
+    "babel-preset-es2015-node5": "^1.1.2",
+    "babel-preset-es2015-webpack": "^6.3.13",
     "babel-preset-react": "^6.3.13",
-    "babel-preset-react-hmre": "^1.0.1",
-    "babel-register": "^6.4.3",
-    "base-error": "^2.0.5",
+    "babel-preset-react-hmre": "^1.1.0",
+    "babel-preset-stage-1": "^6.3.13",
+    "babel-register": "^6.5.2",
     "css-loader": "^0.23.1",
-    "eslint": "^1.10.3",
-    "eslint-config-airbnb": "^3.1.0",
-    "eslint-loader": "^1.2.0",
-    "eslint-plugin-react": "^3.15.0",
+    "eslint": "^2.2.0",
+    "eslint-config-airbnb": "^6.0.2",
     "file-loader": "^0.8.5",
-    "isomorphic-style-loader": "^0.0.8",
+    "isomorphic-style-loader": "0.0.10",
     "json-loader": "^0.5.4",
-    "lodash.mergewith": "^4.0.1",
-    "mime-db": "^1.21.0",
+    "lodash.mergewith": "^4.0.4",
     "mocha": "^2.4.5",
-    "postcss-loader": "^0.8.0",
-    "react-transform-hmr": "^1.0.1",
-    "redux-devtools": "^3.0.1",
-    "supertest": "^1.1.0",
+    "parallelshell": "^2.0.0",
     "url-loader": "^0.5.7",
-    "webpack": "^1.12.12",
-    "webpack-dev-server": "^1.14.0"
+    "webpack": "^2.0.7-beta",
+    "webpack-dev-server": "^2.0.0-beta"
   },
   "peerDependencies": {
-    "history": "^1.17",
-    "react": "^0.14.6",
-    "react-dom": "^0.14.6",
-    "react-router": "^1.0",
-    "jsonwebtoken": "^5.5.4",
-    "koa": "^1.1.2",
-    "koa-bodyparser": "^2.0.1",
-    "koa-jwt": "^1.1.1",
-    "koa-router": "^5.3.0",
-    "koa-static": "^2.0.0",
-    "koa-validate": "^0.2.11",
-    "react-redux": "^4.0.6",
-    "redux": "^3.0.5",
-    "redux-simple-router": "^2.0.2",
-    "redux-thunk": "^1.0.3",
-    "superagent": "^1.7.0",
-    "isomorphic-style-loader": "^0.0.8"
+    "react": "^0.14.7",
+    "react-router": "^2.0.0",
+    "react-router-redux": "^4.0.0",
+    "redux": "^3.2.1",
+    "isomorphic-style-loader": "0.0.8"
   },
   "dependencies": {
-    "jsonwebtoken": "^5.5.4",
     "koa": "^1.1.2",
-    "koa-bodyparser": "^2.0.1",
-    "koa-jwt": "^1.1.1",
-    "koa-router": "^5.3.0",
     "koa-static": "^2.0.0",
-    "koa-validate": "^0.2.11",
-    "react-pure-render": "^1.0.2",
-    "react-redux": "^4.0.6",
-    "redux": "^3.0.5",
-    "redux-simple-router": "^2.0.2",
-    "redux-thunk": "^1.0.3",
-    "superagent": "^1.7.0"
+    "redux-thunk": "^1.0.3"
   }
 }
