@@ -1,8 +1,10 @@
 import { renderToString } from 'react-dom/server';
 import { Provider } from 'react-redux';
 import { RouterContext } from 'react-router';
-import appConfig from '../app_descriptor/app';
-import CSSProvider from '../shared/components/CSSProvider';
+
+
+import appConfig from '../../app_descriptor/app';
+import CSSProvider from '../../shared/components/CSSProvider';
 
 function renderFullPage(html, serializedState, css) {
     return `
@@ -26,13 +28,10 @@ function renderFullPage(html, serializedState, css) {
 function renderBody(store, renderProps, stateSerializer) {
     const css = [];
     const insertCss = (styles) => css.push(styles._getCss());
-    const RootComponent = appConfig.rootComponent;
     const html = renderToString(
         <Provider store={store}>
             <CSSProvider insertCss={insertCss}>
-                <RootComponent>
-                    <RouterContext {...renderProps} />
-                </RootComponent>
+                <RouterContext {...renderProps} />
             </CSSProvider>
         </Provider>
     );
